+++ docs/asset-manifest.json
@@ -1,8 +1,8 @@
 {
   "files": {
     "main.css": "/awesome-crossplane-providers/static/css/main.37f05155.chunk.css",
-    "main.js": "/awesome-crossplane-providers/static/js/main.9ce8804a.chunk.js",
-    "main.js.map": "/awesome-crossplane-providers/static/js/main.9ce8804a.chunk.js.map",
+    "main.js": "/awesome-crossplane-providers/static/js/main.f02356e3.chunk.js",
+    "main.js.map": "/awesome-crossplane-providers/static/js/main.f02356e3.chunk.js.map",
     "runtime-main.js": "/awesome-crossplane-providers/static/js/runtime-main.321ed67a.js",
     "runtime-main.js.map": "/awesome-crossplane-providers/static/js/runtime-main.321ed67a.js.map",
     "static/js/2.72c18a43.chunk.js": "/awesome-crossplane-providers/static/js/2.72c18a43.chunk.js",
@@ -24,6 +24,6 @@
     "static/js/runtime-main.321ed67a.js",
     "static/js/3.594f78f9.chunk.js",
     "static/css/main.37f05155.chunk.css",
-    "static/js/main.9ce8804a.chunk.js"
+    "static/js/main.f02356e3.chunk.js"
   ]
 }
