+++ docs/static/js/main.f02356e3.chunk.js
@@ -1,0 +1,2 @@
+(this["webpackJsonpmaterial-table-demo"]=this["webpackJsonpmaterial-table-demo"]||[]).push([[0],{363:function(e,t,r){},364:function(e,t,r){},433:function(e,t,r){"use strict";r.r(t);var a=r(0),c=r.n(a),n=r(15),j=r.n(n),b=(r(363),r(364),r(224)),f=r.n(b),O=r(19),o=r(335),i=r.n(o),u=r(345),s=r.n(u),d=r(336),l=r.n(d),m=r(343),x=r.n(m),w=r(231),R=r.n(w),h=r(230),p=r.n(h),g=r(337),P=r.n(g),S=r(338),v=r.n(S),C=r(340),k=r.n(C),A=r(341),N=r.n(A),Y=r(342),y=r.n(Y),E=r(346),F=r.n(E),B=r(339),D=r.n(B),J=r(344),M=r.n(J),I=r(347),L=r.n(I),T=r(26),V={Add:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(i.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Check:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(l.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Clear:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(p.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Delete:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(P.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),DetailPanel:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(R.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Edit:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(v.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Export:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(D.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Filter:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(k.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),FirstPage:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(N.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),LastPage:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(y.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),NextPage:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(R.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),PreviousPage:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(x.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),ResetSearch:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(p.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),Search:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(M.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),SortArrow:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(s.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),ThirdStateCheck:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(F.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))})),ViewColumn:Object(a.forwardRef)((function(e,t){return Object(T.jsx)(L.a,Object(O.a)(Object(O.a)({},e),{},{ref:t}))}))},q={data:[{name:"Mohammad",surname:"Faisal",birthYear:1995},{name:"Nayeem Raihan ",surname:"Shuvo",birthYear:1994}],columns:[{title:"Name",field:"name"},{title:"Surname",field:"surname"},{title:"Birth Year",field:"birthYear",type:"numeric"}]},z=function(){return Object(T.jsx)(f.a,{title:"Awesome Crossplane Providers",icons:V,columns:q.columns,data:q.data,options:{sorting:!0,filtering:!0}})};var G=function(){return Object(T.jsx)("div",{className:"App",children:Object(T.jsx)(z,{})})};j.a.render(Object(T.jsx)(c.a.StrictMode,{children:Object(T.jsx)(G,{})}),document.getElementById("root"))}},[[433,1,3]]]);
+//# sourceMappingURL=main.f02356e3.chunk.js.map
